Added fix player 1 mode bug
