+++ src/reducers/mastermind.js
@@ -1,3 +1,5 @@
+import Table from '../logic/table'
+
 const mastermindhistoryinitialState = []
 const mastermindinitialState = {
     CodemakerHist: [],
@@ -7,13 +9,14 @@ const mastermindinitialState = {
     CodemakerTarget: 0,
     CodebreakerTarget: 0,
     isPlayer1turn: true,
-    isOneplayer: false,
+    isOneplayer: true,
     Turn: -1,
     Codebreakerchecked: ["", "", "", ""],
     Codemakerchecked: ["", "", "", ""],
     Secret: [],
     Players: -1,
 }
+const table = Table();
 
 const mastermindhistory = (state = mastermindhistoryinitialState, action) => {
     const { table, type } = action;
@@ -34,22 +37,19 @@ const mastermindhistory = (state = mastermindhistoryinitialState, action) => {
 
 const mastermind = (state = mastermindinitialState, action) => {
     const { actionstate, color, type } = action;
-    const { Codebreakerchecked, CodebreakerHist, CodebreakerTarget, Codemakerchecked, CodemakerHist, CodemakerTarget, Secret } = state;
+    const { isOneplayer, Codebreakerchecked, CodebreakerHist, CodebreakerTarget, Codemakerchecked, CodemakerHist, CodemakerTarget, Secret } = state;
+
     switch (type) {
         case 'UPDATE': {
-            
             return (state != actionstate) ? { ...state, ...actionstate } : { ...state }
         }
         case 'tap_codemakersecret': {
-            return {
-                ...state, Secret: [...Secret, color]
-            }
+            return { ...state, Secret: [...Secret, color] }
         }
         case 'tap_codebreakerchecked': {
             const next = {
                 Codebreakerchecked: [
-                    ...Codebreakerchecked.slice(0, (CodebreakerTarget % 4)),
-                    color, ...Codebreakerchecked.slice((CodebreakerTarget % 4) + 1)]
+                    ...Codebreakerchecked.slice(0, (CodebreakerTarget % 4)), color, ...Codebreakerchecked.slice((CodebreakerTarget % 4) + 1)]
             };
             return {
                 ...state, ...next
@@ -57,22 +57,19 @@ const mastermind = (state = mastermindinitialState, action) => {
         }
         case 'tap_codemakerchecked': {
             return {
-                ...state, Codemakerchecked: [...Codemakerchecked.slice(0,
-                    (CodemakerTarget % 4)),
-                    color,
-                ...Codemakerchecked.slice((CodemakerTarget % 4) + 1)]
+                ...state, Codemakerchecked: [...Codemakerchecked.slice(0, (CodemakerTarget % 4)), color, ...Codemakerchecked.slice((CodemakerTarget % 4) + 1)]
             }
         }
         case 'endCodebreakerTurn': {
-            
             return {
                 ...state, ...actionstate, CodebreakerHist: [...CodebreakerHist, Codebreakerchecked], Codebreakerchecked: ["", "", "", ""]
             }
         }
         case 'endCodemakerTurn': {
-            
+            debugger;
+            const res = (isOneplayer) ? [...CodemakerHist, table.getCodemakerplay(CodebreakerHist[CodebreakerHist.length - 1], Secret)] : [...CodemakerHist, Codemakerchecked]
             return {
-                ...state, ...actionstate, CodemakerHist: [...CodemakerHist, Codemakerchecked], Codemakerchecked: ["", "", "", ""]
+                ...state, ...actionstate, CodemakerHist: res, Codemakerchecked: ["", "", "", ""]
             }
         }
         default: {
